requirements.txt
appdirs==1.4.4
autopep8==2.0.4
customtkinter==5.2.2
darkdetect==0.8.0
docopt==0.6.2
Mako==1.3.2
MarkupSafe==2.1.5
num2words==0.5.13
packaging==23.2
Pillow==10.1.0
pycodestyle==2.11.1
pygame==2.5.2
pygubu==0.34
pygubu-designer==0.38
PyMySQL==1.1.0
pyperclip==1.8.2
screeninfo==0.8.1
